==> src/exp_colectivo/M09_US25_ST100/5_FinalModel_M09_US25_ST100.r <==
#Necesita para correr en Google Cloud
# 128 GB de memoria RAM
# 256 GB de espacio en el disco local
#   8 vCPU

# ZZ final que necesita de UNDERSAMPLING

#limpio la memoria
rm( list=ls() )  #remove all objects
gc()             #garbage collection

require("data.table")

require("lightgbm")

semillas <- c(102191, 869587, 417227, 290923, 66617,66637,66647, 500009, 600011, 700001)

#Parametros del script
PARAM  <- list()
PARAM$experimento  <- "ZZ_M09_US25_ST100" 
PARAM$exp_input  <- "HT9420_M09_US25_ST100" 

PARAM$modelos  <- 1

for (q in semillas ){
  
  PARAM$modelos  <- 1
  # FIN Parametros del script
  
  ksemilla  <- q
  
  #------------------------------------------------------------------------------
  options(error = function() { 
    traceback(20); 
    options(error = NULL); 
    stop("exiting after script error") 
  })
  #------------------------------------------------------------------------------
  #------------------------------------------------------------------------------
  #Aqui empieza el programa
  
  base_dir <- "~/buckets/b1/"
  
  #creo la carpeta donde va el experimento
  dir.create( paste0( base_dir, "exp/", PARAM$experimento, "/"), showWarnings = FALSE )
  setwd(paste0( base_dir, "exp/", PARAM$experimento, "/"))   #Establezco el Working Directory DEL EXPERIMENTO
  
  #leo la salida de la optimizaciob bayesiana
  arch_log  <- paste0( base_dir, "exp/", PARAM$exp_input, "/BO_log.txt" )
  tb_log  <- fread( arch_log )
  setorder( tb_log, -ganancia )
  
  #leo el nombre del expermento de la Training Strategy
  arch_TS  <- paste0( base_dir, "exp/", PARAM$exp_input, "/TrainingStrategy.txt" )
  TS  <- readLines( arch_TS, warn=FALSE )
  
  #leo el dataset donde voy a entrenar el modelo final
  arch_dataset  <- paste0( base_dir, "exp/", TS, "/dataset_train_final.csv.gz" )
  dataset  <- fread( arch_dataset )
  
  # Filtro los meses porque Training Stretegy tiene dos meses de mas
  unique_foto_mes <- unique(dataset$foto_mes)
  unique_foto_mes_sorted <- sort(unique_foto_mes)
  meses_a_quedarse <- unique_foto_mes_sorted[3:length(unique_foto_mes_sorted)]
  dataset <- dataset[foto_mes %in% meses_a_quedarse, ]
  
  
  #leo el dataset donde voy a aplicar el modelo final
  arch_future  <- paste0( base_dir, "exp/", TS, "/dataset_future.csv.gz" )
  dfuture <- fread( arch_future )
  
  
  #defino la clase binaria
  dataset[ , clase01 := ifelse( clase_ternaria %in% c("BAJA+1","BAJA+2"), 1, 0 )  ]
  
  campos_buenos  <- setdiff( colnames(dataset), c( "clase_ternaria", "clase01") )
  
  
  #genero un modelo para cada uno de las modelos_qty MEJORES iteraciones de la Bayesian Optimization
  for( i in  1:PARAM$modelos )
  {
    parametros  <- as.list( copy( tb_log[ i ] ) )
    iteracion_bayesiana  <- parametros$iteracion_bayesiana
    
    arch_modelo  <- paste0( "modelo_" ,
                            sprintf( "%02d", i ),
                            "_",
                            sprintf( "%03d", iteracion_bayesiana ),
                            "_",
                            sprintf( "%03d",ksemilla),
                            ".model" )
    
    
    #creo CADA VEZ el dataset de lightgbm
    dtrain  <- lgb.Dataset( data=    data.matrix( dataset[ , campos_buenos, with=FALSE] ),
                            label=   dataset[ , clase01],
                            weight=  dataset[ , ifelse( clase_ternaria %in% c("BAJA+2"), 1.0000001, 1.0)],
                            free_raw_data= FALSE
    )
    
    ganancia  <- parametros$ganancia
    
    #elimino los parametros que no son de lightgbm
    parametros$experimento  <- NULL
    parametros$cols         <- NULL
    parametros$rows         <- NULL
    parametros$fecha        <- NULL
    parametros$prob_corte   <- NULL
    parametros$estimulos    <- NULL
    parametros$ganancia     <- NULL
    parametros$iteracion_bayesiana  <- NULL
    
    if( ! ("leaf_size_log" %in% names(parametros) ) )  stop( "El Hyperparameter Tuning debe tener en BO_log.txt  el pseudo hiperparametro  lead_size_log.\n" )
    if( ! ("coverage" %in% names(parametros) ) ) stop( "El Hyperparameter Tuning debe tener en BO_log.txt  el pseudo hiperparametro  coverage.\n" )
    
    #Primero defino el tamaño de las hojas
    parametros$min_data_in_leaf  <- pmax( 1,  round( nrow(dtrain) / ( 2.0 ^ parametros$leaf_size_log ))  )
    #Luego la cantidad de hojas en funcion del valor anterior, el coverage, y la cantidad de registros
    parametros$num_leaves  <-  pmin( 131072, pmax( 2,  round( parametros$coverage * nrow( dtrain ) / parametros$min_data_in_leaf ) ) )
    cat( "min_data_in_leaf:", parametros$min_data_in_leaf,  ",  num_leaves:", parametros$num_leaves, "\n" )
    
    #ya no me hacen falta
    parametros$leaf_size_log  <- NULL
    parametros$coverage  <- NULL
    
    #Utilizo la semilla definida en este script
    parametros$seed  <- ksemilla
    
    #genero el modelo entrenando en los datos finales
    set.seed( parametros$seed )
    modelo_final  <- lightgbm( data= dtrain,
                               param=  parametros,
                               verbose= -100 )
    
    #grabo el modelo, achivo .model
    lgb.save( modelo_final,
              file= arch_modelo )
    
    #creo y grabo la importancia de variables
    tb_importancia  <- as.data.table( lgb.importance( modelo_final ) )
    fwrite( tb_importancia,
            file= paste0( "impo_", 
                          sprintf( "%02d", i ),
                          "_",
                          sprintf( "%03d", iteracion_bayesiana ),
                          "_",
                          sprintf( "%03d",ksemilla),
                          ".txt" ),
            sep= "\t" )
    
    
    #genero la prediccion, Scoring
    prediccion  <- predict( modelo_final,
                            data.matrix( dfuture[ , campos_buenos, with=FALSE ] ) )
    
    tb_prediccion  <- dfuture[  , list( numero_de_cliente, foto_mes ) ]
    tb_prediccion[ , prob := prediccion ]
    
    
    nom_pred  <- paste0( "pred_",
                         sprintf( "%02d", i ),
                         "_",
                         sprintf( "%03d", iteracion_bayesiana),
                         "_",
                         sprintf( "%03d",ksemilla),
                         ".csv"  )
    
    fwrite( tb_prediccion,
            file= nom_pred,
            sep= "\t" )
    
    
    #genero los archivos para Kaggle
    #cortes  <- seq( from=  7000,
    #                to=   11000,
    #                by=     500 )
    
    cortes <- c(9000)
    
    
    setorder( tb_prediccion, -prob )
    
    for( corte in cortes )
    {
      tb_prediccion[  , Predicted := 0L ]
      tb_prediccion[ 1:corte, Predicted := 1L ]
      
      nom_submit  <- paste0( PARAM$submit, 
                             "_",
                             sprintf( "%02d", i ),
                             "_",
                             sprintf( "%03d", iteracion_bayesiana ),
                             "_",
                             sprintf( "%05d", corte ),
                             "_",
                             sprintf( "%03d",ksemilla),
                             ".csv" )
      
      fwrite(  tb_prediccion[ , list( numero_de_cliente, Predicted ) ],
               file= nom_submit,
               sep= "," )
      
    }
    
    
    #borro y limpio la memoria para la vuelta siguiente del for
    rm( tb_prediccion )
    rm( tb_importancia )
    rm( modelo_final)
    rm( parametros )
    rm( dtrain )
    gc()
  }
  
}
